pyproject.toml
[project]
name = "AutoPress"
version = "0.1.0"
description = "AutoPress"
authors = [
    {name = "Marc", email = "kastholm1995@gmail.com"}
]
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.5",
    "openai==1.109.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]